InvestmentAdvisory frontend backend all done minor changes and some analysis to be added later

+++ ReactApp2.Server/Models/Investment.cs
@@ -5,7 +5,7 @@
         public int InvestmentId { get; set; }
         public int PortfolioId { get; set; }
         public int AssetId { get; set; }
-        public int PurchasePrice { get; set; }
+        public double PurchasePrice { get; set; }
         public int Quantity { get; set; }
     }
 }
